Added tabs and changed minute integer input to time POSIXlt input.
Tabs created for the different predictors.
Slider input for minute on customer predictor changed to use POSIXlt
datetime objects instead of integers so that direct time HH:MM is
inputted by the user.
Fixed the x-axis on the output plot to display the correct time HH:MM
converted from the minute value.