Added support for user login and sign up.
Created observe even reactive functions for both the login and sign up
buttons. Login event checks the encrypted username and password
submitted by the user matches an entry in the users dataset. Sign up
event checks that the encrypted username submitted by the user is not
already in the dataset and the submitted password is valid, then adds
the new entry to the dataset.